Fix race condition with early CFRunLoop termination.
This commit fixes a race condition that occurs when the CFRunLoop
terminates early, allowing its hosting Goroutine to exit, causing the
underlying thread to be terminated, causing the CFRunLoop to be
deallocated, and thus causing a segmentation fault when CFRunLoopStop is
finally called.

There are a number of reasons that the CFRunLoop can terminate early,
but one reason for termination is the removal of all event sources,
which obviously happens in the stop function with the removal of the
FSEventStream. Whether or not the CFRunLoop terminates quickly enough to
trigger the issue is a matter of chance.

The reason that the underlying thread is terminated is because
runtime.LockOSThread is called without a corresponding
runtime.UnlockOSThread call, in which case the runtime package
explicitly states that the underlying thread is terminated.

The problem is that CFRunLoop instances returned by CFRunLoopGetCurrent
follow "The Get Rule," meaning that the caller has no guarantees about
their lifetime unless they are manually retained/released. It turns out
that thread termination is exactly the time when CFRunLoops are
finalized/deallocated. If the locked thread terminates before
CFRunLoopStop is called, a segmentation fault can occur.

To fix this, the CFRunLoop is now manually retained/released.

In the future, one might also want to ensure that the underlying thread
of the CFRunLoop-hosting Goroutine is released, though this would
require an additional synchronization mechanism to ensure that
CFRunLoopStop is called before that Goroutine returns (to avoid a race
condition where the thread/CFRunLoop is reused and the CFRunLoopStop
call is incorrectly applied to that instance).