Fixed allocator specification for Go 1.11+.
This commit fixes the allocator specification to CFUUIDCreateString,
CFStringCreateCopy, and CFStringCreateWithCString. This was necessary
for Go 1.11+ due to the changes with how pointers are handled when
calling into Objective-C (which sometimes uses pointers for storage
rather than, well, you know, *pointing*).

This commit intentionally avoids fixing the wrap_deprecated.go file
because (a) it's not necessary since it's only used on Go 1.9 and
earlier where this fix is not required and (b) this fork should die soon
once the official repository fixes the issues that it patches.