Add workaround for collision with upstream package.
Docker Compose is now using the upstream fsevents module, which results
in a duplicate symbol error with the exported C callback when building
Mutagen Compose. To fix this, we'll export this module's callback with a
different name. Note that this fix isn't just about avoiding the
collision, it's also about properly routing the callback. Fortunately,
Go's symbol namespacing avoids any Go-Go collisions.

Signed-off-by: Jacob Howard <jacob@mutagen.io>